Remove interactive mode
This is part of KISS. After much struggle with trying to make
interactive mode work with other inputs, i realised it sort of breaks
the Unix philosophy by trying to be two tools, if context building is
needed, it should be done through other means.

In future, we should add the option to print the context with the answer
so that we can then pass in this context again for reuse instead of
keeping context as an internal function.