Added terminal session caching
Adds the ability to cache messages in the terminal session if you add
teh following to .zshrc or .bashrc:

```sh
export CGIP_SESSION_NAME=$(cgip session -i)
```

The reason we had to go this route with te export in the rc is because
you cant really persist a value within a session from a rust program
in any other way.

In the current implementation as of this commit `session -i` simply
generates a uuid which can then be used to identify the session when
cgip is run. This sessionid is used as the key in the cache,  which is
simply a json file stored in ~/.cache/cgip/sessionid